+++ provider/cmd/pulumi-resource-rabbitmq/main.go
@@ -12,16 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:generate go run ./generate.go
+
 package main
 
 import (
-	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfgen"
+	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfbridge"
 
-	rabbitmq "github.com/pulumi/pulumi-rabbitmq"
-	"github.com/pulumi/pulumi-rabbitmq/pkg/version"
+	rabbitmq "github.com/pulumi/pulumi-rabbitmq/provider"
+	"github.com/pulumi/pulumi-rabbitmq/provider/pkg/version"
 )
 
 func main() {
 	// Modify the path to point to the new provider
-	tfgen.Main("rabbitmq", version.Version, rabbitmq.Provider())
+	tfbridge.Main("rabbitmq", version.Version, rabbitmq.Provider(), pulumiSchema)
 }