+++ CHANGELOG.md
@@ -2,7 +2,11 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Upgrade to pulumi-terraform-bridge v1.8.4
+* Upgrade to Pulumi v1.13.1
+* Refactor layout to support Go modules
+
+---
 
 ## 1.2.0 (2020-02-25)
 * Upgrade to v1.3.0 of the RabbitMQ Terraform Provider